backend/pyproject.toml
[tool.poetry]
name = "meeting-synthesis-backend"
version = "0.1.0"
description = "Backend for the Meeting Synthesis application"
authors = ["mplazax <michalplaza9@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.2"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = {extras = ["email"], version = "^2.11.1"}
motor = "^3.7.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
python-dotenv = "^1.0.0"
celery = {extras = ["redis"], version = "^5.5.3"}
redis = "^5.0.1"
future = "^1.0.0"
faster-whisper = "^1.0.0"
numpy = "<2"

ollama = "^0.5.1"
mutagen = "^1.47.0"
pytest = "^8.4.2"
openai = "^1.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.12"
mypy = "^1.16.0"
pytest = "^8.3.5"
pytest-asyncio = "^0.23.8"
httpx = "^0.28.1"
mongomock = "^4.1.2"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
select = ["E", "F", "W", "I", "N", "D", "UP", "ANN", "Q"]
ignore = ["D203", "D212"]
line-length = 88
src = ["app"]

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "venv",
    "backend/.venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "*/migrations/*",
]

[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

plugins = ["pydantic.mypy"]
mypy_path = "app"